AI-CHAT-BOT-FOR-COLLEGE-ENQUIRY-SYSTEM2-1f1980e4e9076456709daee1cbbe8e9fc02b7f38/backend/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-CORS==4.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
mysqlclient==2.2.0
